+++ Develappers.BillomatNet.Tests/InvoiceServiceIntegrationTests.cs
@@ -855,5 +855,119 @@ namespace Develappers.BillomatNet.Tests
 
             var ex = await Assert.ThrowsAsync<NotFoundException>(() => service.EditAsync(inv));
         }
+
+        [Fact]
+        public async Task EditInvoiceItem()
+        {
+            var config = Helpers.GetTestConfiguration();
+            var service = new InvoiceService(config);
+
+            #region Initializing to create
+            var cs = new ClientService(config);
+            var cl = await cs.GetByIdAsync(1506365);
+
+            var articleService = new ArticleService(config);
+            var articles = await articleService.GetByIdAsync(835226);
+
+            var unitService = new UnitService(config);
+            var units = await unitService.GetByIdAsync(articles.UnitId.Value);
+
+            var taxService = new TaxService(config);
+            var taxes = await taxService.GetByIdAsync(articles.TaxId.Value);
+
+            var settingsService = new SettingsService(config);
+            var settings = await settingsService.GetAsync();
+
+            var label = "xUnit Test Object";
+
+            var inv = new Invoice
+            {
+                ClientId = cl.Id,
+                Date = DateTime.Now.Date,
+                Label = label,
+                Quote = 1
+            };
+            #endregion
+
+            var invoiceResult = await service.CreateAsync(inv);
+            Assert.Equal(label, inv.Label);
+
+            var item = new InvoiceItem
+            {
+                InvoiceId = invoiceResult.Id,
+                ArticleId = articles.Id,
+                Unit = units.Name,
+                Quantity = 300,
+                UnitPrice = 1.0f,
+                Title = articles.Title,
+                Description = articles.Description,
+                TaxName = taxes.Name,
+                TaxRate = taxes.Rate,
+            };
+
+            var itemResult = await service.CreateAsync(item);
+            Assert.NotNull(itemResult);
+
+            var editedItem = new InvoiceItem
+            {
+                Id = itemResult.Id,
+                InvoiceId = invoiceResult.Id,
+                ArticleId = articles.Id,
+                Unit = units.Name,
+                Quantity = 300,
+                UnitPrice = 2.0f,
+                Title = articles.Title,
+                Description = articles.Description,
+                TaxName = taxes.Name,
+                TaxRate = taxes.Rate,
+            };
+
+            var editedItemResult = await service.EditAsync(editedItem);
+            Assert.Equal(2.0f, editedItemResult.UnitPrice);
+
+            await service.DeleteAsync(invoiceResult.Id);
+        }
+
+        [Fact]
+        public async Task EditInvoiceItemArgumentException()
+        {
+            var config = Helpers.GetTestConfiguration();
+            var service = new InvoiceService(config);
+
+            var item = new InvoiceItem{};
+
+            var ex = await Assert.ThrowsAsync<ArgumentException>(() => service.EditAsync(item));
+        }
+
+        [Fact]
+        public async Task EditInvoiceItemWhenNotAuthorized()
+        {
+            var config = Helpers.GetTestConfiguration();
+            config.ApiKey = "ajfkjeinodafkejlkdsjklj";
+            var service = new InvoiceService(config);
+
+            var item = new InvoiceItem
+            {
+                Id = 1,
+                InvoiceId = 1
+            };
+
+            var ex = await Assert.ThrowsAsync<NotAuthorizedException>(() => service.EditAsync(item));
+        }
+
+        [Fact]
+        public async Task EditInvoiceItemWhenNotFound()
+        {
+            var config = Helpers.GetTestConfiguration();
+            var service = new InvoiceService(config);
+
+            var item = new InvoiceItem 
+            {
+                Id = 1,
+                InvoiceId = 1
+            };
+
+            var ex = await Assert.ThrowsAsync<NotFoundException>(() => service.EditAsync(item));
+        }
     }
 }